Moved and renamed vision command folder
